Tentatively fix remaining state bugs
The 'pause updating on scroll' changes to MainActivity are quick and
dirty - it's to avoid interrupting the user scrolling through the
delivery list when the list updates and the scroll position resets.

Two approaches I can think of which could be better, but might not be
possible:

1) Tweak the updating not to remove, but instead update, the deliveries.
Might be too complicated, might not even help, but if it works it could
be the cleanest solution

2) Remember the user's scroll position when updating; restore after
update. Should be straightforward _if_ the APIs are available.